Script to help reduce generated box size
This script disables hibernation, removes the sleepimage and deletes
any swapfiles. This reduces uneeded space on disk. Swapfiles will be
re-genereated as needed. If the box is for VMware, it will also make
VMware shrink the disk image via VMware tools. On a Yosemite box
made today this resulted in a reduction from ~14G to ~7G